chore: attempt to improve integration test output
There have been lots of failures recently where tests and timing out on the JVM build.

Because the go test runtime panics we don't get any output, and the tests on CI are configured to fully buffer output.

This change will start streaming output after a set time, so if there is a hang we should be able to get more information.