fix: prevent topic_events being deleted if there is a subscription referencing it
This is because the subscription will lose it’s place if this happens.
If we set it to null, the subscription will start from scratch (unexpected).
If we cascade then we accidentally delete the subscription.
We only delete topic events when trying to fix issues in a cluster, so restricting seems like the safest option to avoid accidental knock on effects.