fix: handle lease failure better
When your laptop goes to sleep we were seeing lots of runner restarts using a cancelled lease context.

It is hard to reproduce reliably, but this should work around it for now.

I could not reproduce the issue by manually cancelling the lease, it only happens when the computer hibernates.

fixes: #3177