feat: store artifacts in OCI registry
This removes the artefacts table and instead stores all binaries in a
container registry. This means that FTL now requires a registry to operate,
and in dev mode it is spun up automatically in a similar manner to postgres.

Uploads are done via the controller, and the runner pulls content directly
from the registry.