feat: Create psql DB in the CF provisioner (#3334)
Connects to the freshly created psql instance, and creates a database if
it does not exist.

This closes https://github.com/TBD54566975/ftl/issues/3117 with
https://github.com/tbdeng/ftl-aws/pull/35

We should refactor the DB creation from the success step to be part of
the actual provisioning flow. Thicket for this:
https://github.com/TBD54566975/ftl/issues/3333

We should also not use the root credentials when connecting to the DB
from a module. However, we should change this when looking at DB
migrations, when we won't expect modules to execute DDL anymore.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>